pyproject.toml
[tool.poetry]
package-mode = false
name = "icfpc-2024"
version = "1.0.0"
description = "ICFPC-2024: Snakes, Monkeys, and Two Smoking Lambdas"
authors = ["Oleg Godovykh <oleg.godovykh@gmail.com>", "Victor Klochikhin <vintlucky777@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
flask-cors = "^4.0.1"
debugpy = "^1.8.2"
numpy = "^2.0.0"
terminaltables = "^3.1.10"
slackclient = "^2.9.4"
python-dotenv = "^1.0.1"
python-dateutil = "^2.9.0.post0"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
